utils functions moved to misc
