+++ resources/lib/items/synced.py
@@ -1,9 +1,8 @@
-#!/usr/bin/python
 # -*- coding: utf-8 -*-
 
 """Defines the SyncedItem class."""
 
-from resources.lib.utils import getstring
+from resources.lib.misc import getstring
 
 
 class SyncedItem(dict):
