+++ resources/lib/utils.py
@@ -14,10 +14,15 @@ import xbmc
 import xbmcgui
 import xbmcvfs
 
+from resources.lib.misc import is_season
+from resources.lib.misc import re_search
+from resources.lib.misc import getstring
+from resources.lib.misc import skip_filter
+from resources.lib.misc import notification
+from resources.lib.misc import SKIP_STRINGS
+
 from resources import ADDON
 from resources import ADDON_ID
-from resources import ADDON_NAME
-from resources import ADDON_PATH
 from resources import RECURSION_LIMIT
 from resources import USING_CUSTOM_MANAGED_FOLDER
 
@@ -33,60 +38,6 @@ else:
     MANAGED_FOLDER = xbmcvfs.translatePath(
         'special://userdata/addon_data/{}/'.format(ADDON_ID)
     )
-
-
-def notification(message, time=3000, icon=join(ADDON_PATH, 'ntf_icon.png')):
-    """Provide a shorthand for xbmc builtin notification with addon name."""
-    xbmcgui.Dialog().notification(
-        ADDON_NAME,
-        str(message),
-        icon,
-        time,
-        True
-    )
-
-
-def tojs(data, filename):
-    """Function to create a json file."""
-    try:
-        with open(join(expanduser('~/'), filename) + '.json', 'a+') as f:
-            f.write(str(json.dumps(data, indent=4, sort_keys=True)))
-            f.close()
-    except AttributeError:
-        pass
-
-
-def getstring(string_id):
-    """Shortcut function to return string from String ID."""
-    return str(ADDON.getLocalizedString(string_id))
-
-
-def title_with_color(label, year=None, color='mediumslateblue'):
-    """Create a string to use in title Dialog().select."""
-    # COLORS: https://github.com/xbmc/xbmc/blob/master/system/colors.xml
-    # TODO: this function can be better, maybe led generic,
-    # now, this func add color and year to movie title,
-    # and any of this actions can be splited
-    if year:
-        return str('[COLOR %s][B]%s (%s)[/B][/COLOR]' % (color, label, year))
-    return str('[COLOR %s][B]%s[/B][/COLOR]' % (color, label))
-
-
-def color(string, colorname='mediumslateblue'):
-    """
-    Return string formated with a selected color.
-    lawngreen
-    mediumseagreen
-    """
-    # COLORS: https://github.com/xbmc/xbmc/blob/master/system/colors.xml
-    return str('[COLOR %s]%s[/COLOR]' % (colorname, string))
-
-
-def bold(string):
-    """
-    Return string formated with a bold.
-    """
-    return str('[B]%s[/B]' % (string))
 
 
 def check_managed_folder():
@@ -184,57 +135,6 @@ def jsonrpc_getdirectory(_path):
         )['result']['files']
     except KeyError:
         log_msg('KeyError in return of JSONRPC.')
-
-
-def videolibrary(method, database='video', path=None):
-    """A dedicated method to performe jsonrpc VideoLibrary.Scan or VideoLibrary."""
-    command = {
-        'scan': 'CleanLibrary(%s)' % database,
-        'clean': 'UpdateLibrary(%s)' % database
-    }
-    xbmc.executebuiltin(command[method])
-
-
-SKIP_STRINGS = [
-    'resumo',
-    'suggested',
-    'extras',
-    'trailer',
-    r'\#(?:\d{1,5}\.\d{1,5}|SP)',
-]
-
-
-def re_search(string, tosearch=None):
-    """Function check if string exist with re."""
-    tosearch = tosearch if isinstance(tosearch, list) else [tosearch]
-    return bool(
-        any(
-            re.search(
-                rgx,
-                string,
-                re.I
-            ) for rgx in tosearch
-        )
-    )
-
-
-def skip_filter(contents_json, _key, toskip):
-    """Function to iterate jsons in a list and filter by key with re."""
-    try:
-        for item in contents_json:
-            if not bool(any(re.search(rgx, item[_key], re.I) for rgx in toskip)):
-                yield item
-    except TypeError:
-        yield None
-
-
-def is_season(string):
-    return bool(
-        re_search(
-            string,
-            ['season', 'temporada', r'S\d{1,4}']
-        )
-    )
 
 
 def list_reorder(contents_json, showtitle, sync_type=False):
